Exercises from week #7 added
Creation of Firebase Project, adding Google-services.json to configure with Firebase. Implemented an Authentication System based on Google Firebase Authentication. Firebase UI was used for a streamlined setup. Authentication platforms that are integrated are Google Sign-in, Email Sign-in, Facebook Integration wwith Auth0 and Guest user. Functionality of the app is different from when you have created an account to when you continue as a guest. A Navigation Drawer has been implemented, and here the visibility of app options are again based on the user’s authentication status.

Would like to implement navigation from the navigation drawer i.e when you click "add event" you will go to the add event fragment - same way as you do from the bottomnavigation.